Summary of changes
server.js - Added a datastructure purge until I find a better solution.
MyFileWriter - A new class to store all user transactions in a file.
common - A new directory for all the common stuff
RecvieveSMS - Name lookup in the contact book for a given number.
WebViewActivity - Added a locationListener separtely for GPS tracking.
BestLocationFinder - Changed the constructor to not accept the frequency parameter.